:sparkles: feat: Batch jobs that are run via CLI interaction mode are managed by the running job registry
