+++ tofhir-engine/src/main/scala/io/tofhir/engine/cli/CommandExecutionContext.scala
@@ -7,10 +7,5 @@ import scala.concurrent.Future
 
 case class CommandExecutionContext(toFhirEngine: ToFhirEngine,
                                    fhirMappingJob: Option[FhirMappingJob] = None,
-                                   mappingNameUrlMap: Map[String, String] = Map.empty,
-                                   runningStatus: Option[(Option[FhirMappingTask], Future[Unit])] = None) {
-
-  def withStatus(newRunningStatus: Option[(Option[FhirMappingTask], Future[Unit])]): CommandExecutionContext = {
-    this.copy(runningStatus = newRunningStatus)
-  }
+                                   mappingNameUrlMap: Map[String, String] = Map.empty) {
 }
