+++ tofhir-engine/src/main/scala/io/tofhir/engine/cli/CommandFactory.scala
@@ -5,7 +5,6 @@ object CommandFactory {
   def apply(command: String): Command = {
     command match {
       case "help" => new Help()
-      case "info" => new Info()
       case "load" => new Load()
       case "extract-redcap-schemas" => new ExtractRedCapSchemas()
       case "reload" => new Reload()
