+++ tofhir-engine/src/main/scala/io/tofhir/engine/cli/CommandLineInterface.scala
@@ -68,15 +68,6 @@ object CommandLineInterface {
       val commandName = Try(args.head).getOrElse("")
       val commandArgs = Try(args.tail).getOrElse(Seq.empty[String])
       commandExecutionContext = CommandFactory.apply(commandName).execute(commandArgs, commandExecutionContext)
-      commandExecutionContext.runningStatus.foreach(_._2.onComplete {
-        case Success(_) =>
-          println("Job/Task running completed successfully for the command.")
-          commandExecutionContext = commandExecutionContext.withStatus(Option.empty)
-        case Failure(exception) =>
-          println("Problem during execution of the command!")
-          exception.printStackTrace()
-          commandExecutionContext = commandExecutionContext.withStatus(Option.empty)
-      })
       print("\n$ ")
     }
   }
