+++ tofhir-engine/src/main/scala/io/tofhir/engine/cli/Help.scala
@@ -3,7 +3,6 @@ package io.tofhir.engine.cli
 class Help extends Command {
   override def execute(args: Seq[String], context: CommandExecutionContext): CommandExecutionContext = {
     println("List of available commands:\n" +
-      "\tinfo - See info about the loaded Mapping Job.\n" +
       "\tload <path> - Load the Mapping Job definition file from the path.\n" +
       "\treload - Reload the mapping definitions from their source into the mapping repository.\n" +
       "\trun [<url>|<name>] - Run the task(s). Without a parameter, all task of the loaded Mapping Job are run. A specific task can be indicated with its name or URL.\n" +
