+++ tofhir-engine/src/main/scala/io/tofhir/engine/cli/Load.scala
@@ -2,7 +2,7 @@ package io.tofhir.engine.cli
 
 import io.tofhir.engine.cli
 import io.tofhir.engine.mapping.IFhirMappingRepository
-import io.tofhir.engine.model.FhirMappingTask
+import io.tofhir.engine.model.{FhirMappingTask, FhirRepositorySinkSettings, FhirSinkSettings, FileSystemSourceSettings, KafkaSourceSettings, SqlSourceSettings}
 import io.tofhir.engine.util.FhirMappingJobFormatter
 import org.json4s.MappingException
 
@@ -22,7 +22,7 @@ class Load extends Command {
         val mappingJob = FhirMappingJobFormatter.readMappingJobFromFile(filePath)
         println("The following FhirMappingJob successfully loaded.")
         val newContext = cli.CommandExecutionContext(context.toFhirEngine, Some(mappingJob), Load.getMappingNameUrlTuples(mappingJob.mappings, context.toFhirEngine.mappingRepo))
-        println(Info.serializeMappingJobToCommandLine(newContext))
+        println(Load.serializeMappingJobToCommandLine(newContext))
         newContext
       } catch {
         case _: FileNotFoundException =>
@@ -41,5 +41,43 @@ object Load {
     tasks.foldLeft(Map.empty[String, String]) { (map, task) => // Convert to tuple (name -> url)
       map + (mappingRepository.getFhirMappingByUrl(task.mappingRef).name -> task.mappingRef)
     }
+  }
+
+  def serializeMappingJobToCommandLine(context: CommandExecutionContext): String = {
+    if (context.fhirMappingJob.isEmpty) {
+      throw new IllegalStateException("!!! I am trying to serialize the MappingJob from the context, but it is not there!!!")
+    }
+    val mj = context.fhirMappingJob.get
+    val sourceSettingsStr = mj.sourceSettings.map {
+      case (sourceAlias, settings: FileSystemSourceSettings) =>
+        s"\tFile System Source Settings for '$sourceAlias':\n" +
+          s"\t\tName: ${settings.name},\n" +
+          s"\t\tSource URI: ${settings.sourceUri},\n" +
+          s"\t\tData Folder Path: ${settings.dataFolderPath}\n"
+      case (sourceAlias, settings: SqlSourceSettings) =>
+        s"\tSql Source Settings for '$sourceAlias':\n" +
+          s"\t\tName: ${settings.name},\n" +
+          s"\t\tSource URI: ${settings.sourceUri},\n" +
+          s"\t\tDatabase URL: ${settings.databaseUrl},\n" +
+          s"\t\tUsername: ${settings.username},\n" +
+          s"\t\tPassword: ********\n"
+      case (sourceAlias, settings: KafkaSourceSettings) =>
+        s"\tKafka Source Settings for '$sourceAlias':\n" +
+          s"\t\tName: ${settings.name},\n" +
+          s"\t\tSource URI: ${settings.sourceUri},\n" +
+          s"\t\tBootstrap servers: ${settings.bootstrapServers}\n"
+      case _ => "\tNo Source Settings\n"
+    }.mkString("\n")
+    val sinkSettingsStr = mj.sinkSettings match {
+      case settings: FhirRepositorySinkSettings =>
+        s"\tFHIR Repository URL: ${settings.fhirRepoUrl}\n"
+      case _: FhirSinkSettings => "\tNo Sink Settings\n"
+    }
+
+    val tasks = context.mappingNameUrlMap
+      .map { case (name, url) => s"\t\t$name -> $url" } // Convert to string
+    val tasksStr = "\tTasks:\n" + tasks.mkString("\n")
+
+    s"Mapping Job ID: ${mj.id}\n" + sourceSettingsStr + sinkSettingsStr + tasksStr
   }
 }
